+++ apps/e2e/src/Controller/TurboController.php
@@ -1,5 +1,14 @@
 <?php
 
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace App\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -12,13 +21,11 @@ use Symfony\UX\Turbo\TurboBundle;
 #[Route('/ux-turbo', name: 'app_ux_turbo_')]
 final class TurboController extends AbstractController
 {
-
     #[Route('/drive', name: 'drive')]
     public function drive(
         #[MapQueryParameter] int $page = 1,
-    ): Response
-    {
-        if ($page === 2) {
+    ): Response {
+        if (2 === $page) {
             return $this->render('ux_turbo/drive_page_2.html.twig', [
                 'current_time' => (new \DateTimeImmutable())->format(\DateTimeInterface::RFC3339_EXTENDED),
             ]);
@@ -28,7 +35,6 @@ final class TurboController extends AbstractController
             'current_time' => (new \DateTimeImmutable())->format(\DateTimeInterface::RFC3339_EXTENDED),
         ]);
     }
-
 
     #[Route('/frame', name: 'frame')]
     public function frame(): Response
@@ -48,6 +54,7 @@ final class TurboController extends AbstractController
         if ($request->isMethod('POST')) {
             if (TurboBundle::STREAM_FORMAT === $request->getPreferredFormat()) {
                 $request->setRequestFormat(TurboBundle::STREAM_FORMAT);
+
                 return $this->render('ux_turbo/stream_response.html.twig');
             }