+++ apps/e2e/src/Form/Type/VideogameAutocompleteType.php
@@ -1,5 +1,14 @@
 <?php
 
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace App\Form\Type;
 
 use Symfony\Component\Form\AbstractType;
@@ -10,7 +19,7 @@ use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 class VideogameAutocompleteType extends AbstractType
 {
     public function __construct(
-        private UrlGeneratorInterface $urlGenerator
+        private UrlGeneratorInterface $urlGenerator,
     ) {
     }