+++ apps/e2e/src/Twig/Components/LiveFruitsPagination.php
@@ -1,5 +1,14 @@
 <?php
 
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace App\Twig\Components;
 
 use App\Repository\FruitRepository;
@@ -19,7 +28,7 @@ final class LiveFruitsPagination
 
     public function __construct(
         private FruitRepository $fruitRepository,
-    )    {
+    ) {
     }
 
     public function getFruits(): array
@@ -44,7 +53,7 @@ final class LiveFruitsPagination
     public function goToPreviousPage(): void
     {
         if ($this->hasPreviousPage()) {
-            $this->page--;
+            --$this->page;
         }
     }
 
@@ -52,7 +61,7 @@ final class LiveFruitsPagination
     public function goToNextPage(): void
     {
         if ($this->hasNextPage()) {
-            $this->page++;
+            ++$this->page;
         }
     }
 }