+++ src/Translator/CHANGELOG.md
@@ -54,6 +54,12 @@
 
     **Note:** This is a breaking change, but the UX Translator component is still experimental.
 
+-   **[BC BREAK]** Refactor `TranslationsDumper` to accept configuration options via `dump()` method parameters, instead of constructor arguments or method calls:
+        - Removed `$dumpDir` and `$dumpTypeScript` constructor arguments
+        - Removed `TranslationsDumper::addIncludedDomain()` and `TranslationsDumper::addExcludedDomain()` methods
+
+    **Note:** This is a breaking change, but the UX Translator component is still experimental.
+
 -   Add configuration `ux_translator.dump_typescript`  to enable/disable TypeScript types dumping,
     default to `true`. Generating TypeScript types is useful when developing,
     but not in production when using the AssetMapper (which does not use these types).